refactor: :bento: Change name of assets
Change name of assets
